Stop partnership with Magikcraft and tweak layout of patners:
Magikcraft stops provideing Japanese servers with CoderDojo.
Each dojo still can access to Magikcraft-hosted servers, but
that is not only for CoderDojos in Japan but for anyone.

To clarify what makes partners and fairly treat all of them,
CoderDojo Japan define partnerships and available terms by this:
cf. https://coderdojo.jp/docs/term-of-partnership